requirements.txt
instructor
openai
dspy-ai
tqdm
polars
python-dotenv
anthropic # for instructor